+++ services/workflows-service/scripts/seed.ts
@@ -3,9 +3,9 @@ import { faker } from '@faker-js/faker';
 import { Business, Customer, EndUser, Prisma, PrismaClient, Project } from '@prisma/client';
 import { Type } from '@sinclair/typebox';
 import { hash } from 'bcrypt';
+import { hashKey } from '../src/customer/api-key/utils';
 import { env } from '../src/env';
 import type { InputJsonValue } from '../src/types';
-import { hashKey } from './../src/customer/api-key/utils';
 import { seedTransactionsAlerts } from './alerts/generate-alerts';
 import { generateTransactions } from './alerts/generate-transactions';
 import { seedBusinessReports } from './business-reports/seed-business-reports';
@@ -938,24 +938,6 @@ async function seed() {
     project1.id,
   );
 
-  await client.$transaction(async () =>
-    endUserIds.map(async (id, index) =>
-      client.endUser.create({
-        /// I tried to fix that so I can run through ajv, currently it doesn't like something in the schema (anyOf  )
-        data: generateEndUser({
-          id,
-          workflow: {
-            workflowDefinitionId: kycManualMachineId,
-            workflowDefinitionVersion: manualMachineVersion,
-            context: await createMockEndUserContextData(id, index + 1),
-            state: DEFAULT_INITIAL_STATE,
-          },
-          projectId: project1.id,
-        }),
-      }),
-    ),
-  );
-
   await client.$transaction(async tx => {
     businessRiskIds.map(async (id, index) => {
       const riskWf = async () => ({
@@ -1026,6 +1008,26 @@ async function seed() {
   //     },
   //   },
   // });
+
+  await client.$transaction(async () =>
+    endUserIds.map(async (id, index) =>
+      client.endUser.create({
+        /// I tried to fix that so I can run through ajv, currently it doesn't like something in the schema (anyOf  )
+        data: generateEndUser({
+          id,
+          workflow: {
+            workflowDefinitionId: kycManualMachineId,
+            workflowDefinitionVersion: manualMachineVersion,
+            context: await createMockEndUserContextData(id, index + 1),
+            state: DEFAULT_INITIAL_STATE,
+          },
+          projectId: project1.id,
+          connectBusinesses: Math.random() > 0.5,
+        }),
+      }),
+    ),
+  );
+
   void client.$disconnect();
 
   console.info('Seeding database with custom seed...');